Tidying up the validation code, now to get happy with windows
