+++ packages/workbox-cli/test/node/inject-manifest-e2e.js
@@ -3,7 +3,7 @@ const fsExtra = require('fs-extra');
 const path = require('path');
 const proxyquire = require('proxyquire');
 
-require('chai').should();
+const expect = require('chai').expect;
 
 describe('Generate SW End-to-End Tests', function() {
   let tmpDirectory;
@@ -73,33 +73,34 @@ describe('Generate SW End-to-End Tests', function() {
     return cli.handleCommand('inject:manifest')
     .then(() => {
       const fileOutput = fs.readFileSync(swDest).toString();
-      fileOutput.should.equal(`someVariables.precache([
+      const regex = new RegExp(`someVariables.precache\\(\\[
   {
     "url": "/images/web-fundamentals-icon192x192.png",
-    "revision": "93ffb20d77327583892ca47f597b77aa"
+    "revision": "\\w*"
   },
   {
     "url": "/index.html",
-    "revision": "24abd5daf6d87c25f40c2b74ee3fbe93"
+    "revision": "\\w*"
   },
   {
     "url": "/page-1.html",
-    "revision": "544658ab25ee8762dc241e8b1c5ed96d"
+    "revision": "\\w*"
   },
   {
     "url": "/page-2.html",
-    "revision": "a3a71ce0b9b43c459cf58bd37e911b74"
+    "revision": "\\w*"
   },
   {
     "url": "/styles/stylesheet-1.css",
-    "revision": "934823cbc67ccf0d67aa2a2eeb798f12"
+    "revision": "\\w*"
   },
   {
     "url": "/styles/stylesheet-2.css",
-    "revision": "884f6853a4fc655e4c2dc0c0f27a227c"
+    "revision": "\\w*"
   }
-]);
-`);
+\\]\\);`);
+      const result = regex.exec(fileOutput);
+      expect(result).to.exist;
     })
     .then(() => {
       // Should be able to handle command with no questions
